+++ src/common/main/java/dev/jolkert/stayawhile/StayAWhile.java
@@ -1,0 +1,11 @@
+package dev.jolkert.stayawhile;
+
+import net.minecraft.world.level.GameRules;
+
+public class StayAWhile
+{
+	public static GameRules.Key<GameRules.IntegerValue> MAX_ITEM_AGE;
+	public static GameRules.Key<GameRules.IntegerValue> MAX_PLAYER_THROWN_ITEM_AGE;
+	public static GameRules.Key<GameRules.IntegerValue> MAX_PLAYER_DEATH_ITEM_AGE;
+	public static GameRules.Key<GameRules.BooleanValue> SCATTER_DEATH_ITEMS;
+}