+++ src/common/main/java/dev/jolkert/stayawhile/duck/ItemEntityDuck.java
@@ -1,0 +1,9 @@
+package dev.jolkert.stayawhile.duck;
+
+public interface ItemEntityDuck
+{
+	boolean stayawhile$isDeathDrop();
+	void stayawhile$setDeathDrop(boolean deathDrop);
+
+	boolean stayawhile$wasThrown();
+}