+++ src/main/java/io/github/jolkert/stayawhile/access/ItemEntityInterface.java
@@ -1,9 +1,0 @@
-package io.github.jolkert.stayawhile.access;
-
-import io.github.jolkert.stayawhile.data.DropType;
-
-public interface ItemEntityInterface
-{
-	void stayAWhile$setDropType(DropType dropType);
-	DropType stayAWhile$getDropType();
-}